fix time ordering issue between moment.js, python, and sqlite3
Moment.js updates cause a timestamp to be like:

 2015-01-02 12:59:00+00:00

python creates timestamps with a "T"

 2015-01-02T12:59:00+00:00

sqlite seems to be sorting by string and not and time value so
the missing "T" is causing things to sort wrong.

This also makes sure the "api" isn't tied to the whims of the
client's uploading to it.